+++ Assets/GameJam/Script/InGame/Item/EntertainmentItem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc65fe9e4dc7e2243a612f9fa64ab19c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
