+++ Assets/GameJam/Script/InGame/Item/GamblingItem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19bc9a0cf362fb2468d0f89d6fd464ec
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
