+++ Assets/GameJam/Script/InGame/Item/GamblingItem/GamblingItemController.cs
@@ -5,19 +5,33 @@ using Tech.C.Interface;
 namespace Tech.C.Item
 {
     /// <summary>
-    /// 娯楽系アイテムの挙動を管理するコントローラー
+    /// ギャンブル系アイテムの挙動を管理するコントローラー
     /// </summary>
-    public class EntertainmentItemController : MonoBehaviour, IFallingItem
+    public class GamblingItemController : MonoBehaviour, IFallingItem
     {
-        [SerializeField] private EntertainmentType entertainmentType;
-        private MoveType moveType;
-
+        [SerializeField] private GamblingType gamblingType;
+        private ItemMoveType moveType;
         [SerializeField] private float fallSpeed = 2f;
-
         private Rigidbody2D rb;
 
+        // Pool参照を保持
         private ItemPool itemPool;
         private int poolIndex;
+
+        /// <summary>
+        /// 外部からタイプをセットする
+        /// </summary>
+        public void SetType(GamblingType type)
+        {
+            gamblingType = type;
+        }
+
+        // Pool参照をセットするメソッド
+        public void SetPool(ItemPool pool, int index)
+        {
+            itemPool = pool;
+            poolIndex = index;
+        }
 
         void Start()
         {
@@ -29,28 +43,15 @@ namespace Tech.C.Item
             Fall();
         }
 
-        // 毎フレームの落下・移動処理
         public void Fall()
         {
-            rb.linearVelocity = new Vector2(0, -fallSpeed);
-        }
-        /// <summary>
-        /// 外部からタイプをセットする
-        /// </summary>
-        public void SetType(EntertainmentType type)
-        {
-            entertainmentType = type;
+            if (rb != null)
+                rb.linearVelocity = new Vector2(0, -fallSpeed);
         }
 
-        // Pool参照をセットするメソッド
-        public void SetPool(ItemPool pool, int index)
-        {
-            itemPool = pool;
-            poolIndex = index;
-        }
 
-        // 共通返却処理
-        private void ReturnToPool()
+        // アイテムをPoolに返却
+        public void ReturnToPool()
         {
             if (itemPool != null)
             {
@@ -77,7 +78,7 @@ namespace Tech.C.Item
         // 弾との衝突判定
         void OnTriggerEnter2D(Collider2D other)
         {
-            if (other.CompareTag("Bullet"))
+            if (other.CompareTag("Wall"))
             {
                 OnCollected();
             }
