+++ Assets/GameJam/Script/InGame/Item/Item.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9da92409eac4dcf4b8098e57add00020
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
