+++ Assets/GameJam/Script/InGame/Item/Item/ItemSpawner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f056711cf02dc4944a7f40547738205e
