--- Assets/GameJam/Script/InGame/Item/MoveType.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f60fe71e7e5921f489355ebc4418ad91
