+++ Assets/GameJam/Script/InGame/Manager/GaugeController.cs
