+++ Assets/GameJam/Script/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1507c673b66e8494bab9f5cf7c9e0121
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
